% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{insert a flag}
\usage{
flag(country, size = 30, squared = FALSE, width = NULL)
}
\arguments{
\item{country}{a character, the \href{https://www.iso.org/obp/ui/#search/code/}{ISO 3166-1-alpha-2 country code}}

\item{size}{an integer, the value of the \code{font-size} CSS parameter in pixels,}

\item{squared}{a logical, if FALSE (the default) use the flags with 4x3 ratio, else 1x1}

\item{width}{a character, the value of the \code{width} CSS parameter}
}
\description{
insert a flag
}
\examples{

flag("FR")

flag("US", squared = TRUE)

flag("US", squared = TRUE, width = "300px")
}
